[Refactoring ⚙️] 모달 호출을 전역에서 하도록 수정
