+++ src/pages/Recommend/Machine/Machine.tsx
@@ -2,7 +2,6 @@ import { useRef, useState } from "react"
 import { useLocation, useNavigate } from "react-router-dom"
 
 import IconButton from "@components/IconButton/IconButton"
-import AlertLoadingModal from "@components/Modal/components/Alert/AlertLoadingModal"
 import ProgressBar from "@components/Progressbar/ProgressBar"
 
 import Footer from "@pages/Recommend/Machine/Footer"
@@ -87,7 +86,6 @@ const Machine = () => {
         postRecommend={postRecommend}
         machines={machines}
       />
-      <AlertLoadingModal />
     </>
   )
 }
